Rename migration_connector to schema_connector
The trait was previously renamed from `MigrationConnector` to
`SchemaConnector` but the file wasn't. This commit fixes that.
